# apps/docs/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mayarin API",
    "version": "2026-08-11",
    "description": "Create fiat-denominated commerce requests, accept supported crypto assets, and settle merchants in stablecoins. Versioned operations live under `/v1`; buyer-facing checkout, invoice, and health routes are intentionally unversioned."
  },
  "servers": [
    {
      "url": "https://api.mayarin.xyz",
      "description": "Mainnet"
    },
    {
      "url": "https://api-testnet.mayarin.xyz",
      "description": "Testnet"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "secretKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Server-side `sk_` key. Grants a subset of the merchant's permissions."
      },
      "publishableOrSecretKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Browser-safe `pk_` key, or a stronger server-side `sk_` key. Identifies the merchant for catalog read and cart checkout; carries no permissions."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              }
            },
            "required": ["code", "message", "retryable"]
          }
        },
        "required": ["error"]
      },
...
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "merchantId": {
            "type": "string"
          },
          "merchant": {
            "$ref": "#/components/schemas/Merchant"
          },
          "number": {
            "type": ["string", "null"]
          },
          "sequence": {
            "type": ["integer", "null"]
          },
          "state": {
            "type": "string",
            "description": "DRAFT, ISSUED, PAID, or VOID."
          },
          "buyer": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "type": ["string", "null"]
              },
              "taxId": {
                "type": ["string", "null"]
              },
              "address": {
                "type": ["string", "null"]
              }
            },
            "required": ["name", "email", "taxId", "address"]
          },
          "currency": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": ["string", "null"]
                },
                "name": {
                  "type": "string"
                },
                "unitPrice": {
                  "$ref": "#/components/schemas/Money"
                },
                "quantity": {
                  "type": "integer"
                }
              },
              "required": ["productId", "name", "unitPrice", "quantity"]
            }
          },
          "total": {
            "$ref": "#/components/schemas/Money"
          },
          "notes": {
            "type": ["string", "null"]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "url": {
            "type": "string"
          },
          "issuedAt": {
            "type": ["string", "null"]
          },
          "dueAt": {
            "type": ["string", "null"]
          },
          "voidedAt": {
            "type": ["string", "null"]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "listed": {
            "type": "boolean",
            "description": "Whether it appears in the public x402 payable index."
          }
        },
        "required": [
          "id",
          "merchantId",
          "merchant",
          "number",
          "sequence",
          "state",
          "buyer",
          "currency",
          "lines",
          "total",
          "notes",
          "metadata",
          "url",
          "issuedAt",
          "dueAt",
          "voidedAt",
          "createdAt",
          "updatedAt",
          "version",
          "listed"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/invoices": {
      "get": {
        "tags": ["Invoices"],
        "operationId": "listInvoices",
        "summary": "List invoices",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "merchantId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoices.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Invoices"],
        "operationId": "createInvoice",
        "summary": "Create a draft invoice",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Stable unique key for this logical write. Keep it unchanged when retrying."
            },
            "required": true,
            "description": "Stable unique key for this logical write. Keep it unchanged when retrying.",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "merchantId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "merchant": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "minLength": 1
                      },
                      "name": {
                        "type": "string",
                        "minLength": 1
                      },
                      "city": {
                        "type": "string",
                        "minLength": 1
                      },
                      "countryCode": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 2
                      },
                      "categoryCode": {
                        "type": "string"
                      }
                    },
                    "required": ["id", "name", "city", "countryCode"]
                  },
                  "buyer": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 255
                      },
                      "email": {
                        "type": "string",
                        "maxLength": 320,
                        "format": "email"
                      },
                      "taxId": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 64
                      },
                      "address": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000
                      }
                    },
                    "required": ["name"]
                  },
                  "currency": {
                    "type": "string",
                    "enum": [
                      "IDR",
                      "USD",
                      "SGD",
                      "THB",
                      "MYR",
                      "USDC",
                      "USDT",
                      "EURC",
                      "ETH",
                      "BTC"
                    ]
                  },
                  "lines": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 255
                        },
                        "unitPrice": {
                          "type": "object",
                          "properties": {
                            "amount": {
                              "type": "string"
                            },
                            "asset": {
                              "type": "string",
                              "enum": [
                                "IDR",
                                "USD",
                                "SGD",
                                "THB",
                                "MYR",
                                "USDC",
                                "USDT",
                                "EURC",
                                "ETH",
                                "BTC"
                              ]
                            }
                          },
                          "required": ["amount", "asset"]
                        },
                        "quantity": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        }
                      },
                      "required": ["name", "unitPrice", "quantity"]
                    },
                    "minItems": 1
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                },
                "required": ["merchantId", "merchant", "buyer", "currency", "lines"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft invoice created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invoices/{id}": {
      "get": {
        "tags": ["Invoices"],
        "operationId": "getInvoice",
        "summary": "Get an invoice",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice with derived paid/outstanding figures.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceView"
                }
              }
            }
          },
          "404": {
            "description": "No invoice with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Invoices"],
        "operationId": "updateInvoice",
        "summary": "Update a draft invoice",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Stable unique key for this logical write. Keep it unchanged when retrying."
            },
            "required": true,
            "description": "Stable unique key for this logical write. Keep it unchanged when retrying.",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "buyer": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 255
                      },
                      "email": {
                        "type": "string",
                        "maxLength": 320,
                        "format": "email"
                      },
                      "taxId": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 64
                      },
                      "address": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000
                      }
                    },
                    "required": ["name"]
                  },
                  "lines": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 255
                        },
                        "unitPrice": {
                          "type": "object",
                          "properties": {
                            "amount": {
                              "type": "string"
                            },
                            "asset": {
                              "type": "string",
                              "enum": [
                                "IDR",
                                "USD",
                                "SGD",
                                "THB",
                                "MYR",
                                "USDC",
                                "USDT",
                                "EURC",
                                "ETH",
                                "BTC"
                              ]
                            }
                          },
                          "required": ["amount", "asset"]
                        },
                        "quantity": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        }
                      },
                      "required": ["name", "unitPrice", "quantity"]
                    },
                    "minItems": 1
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated invoice.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "description": "No invoice with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invoices/{id}/issue": {
      "post": {
        "tags": ["Invoices"],
        "operationId": "issueInvoice",
        "summary": "Issue an invoice",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Stable unique key for this logical write. Keep it unchanged when retrying."
            },
            "required": true,
            "description": "Stable unique key for this logical write. Keep it unchanged when retrying.",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dueAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "prefix": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 16
                  },
                  "includeYear": {
                    "type": "boolean"
                  }
                },
                "required": ["dueAt"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Issued invoice.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "description": "No invoice with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invoices/{id}/void": {
      "post": {
        "tags": ["Invoices"],
        "operationId": "voidInvoice",
        "summary": "Void an invoice",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Stable unique key for this logical write. Keep it unchanged when retrying."
            },
            "required": true,
            "description": "Stable unique key for this logical write. Keep it unchanged when retrying.",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Voided invoice.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "description": "No invoice with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invoices/{id}/list": {
      "post": {
        "tags": ["Invoices"],
        "operationId": "listInvoice",
        "summary": "List an invoice in the public payable index",
        "description": "Opts the invoice into the public x402 payable index, where agents discover and pay it. The invoice stays payable by anyone holding its id either way; listing only decides whether a discovery reader is shown it.",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Listed invoice.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "description": "No invoice with that id, or not this merchant's.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invoices/{id}/unlist": {
      "post": {
        "tags": ["Invoices"],
        "operationId": "unlistInvoice",
        "summary": "Remove an invoice from the public payable index",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Unlisted invoice.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "description": "No invoice with that id, or not this merchant's.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invoices/{id}/checkout": {
      "post": {
        "tags": ["Invoices"],
        "operationId": "checkoutInvoice",
        "summary": "Check out an issued invoice",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Stable unique key for this logical write. Keep it unchanged when retrying."
            },
            "required": true,
            "description": "Stable unique key for this logical write. Keep it unchanged when retrying.",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "payment": {
                    "type": "object",
                    "properties": {
                      "asset": {
                        "type": "string",
                        "enum": [
                          "IDR",
                          "USD",
                          "SGD",
                          "THB",
                          "MYR",
                          "USDC",
                          "USDT",
                          "EURC",
                          "ETH",
                          "BTC"
                        ]
                      },
                      "chain": {
                        "type": "string",
                        "enum": [
                          "base",
                          "base-sepolia",
                          "arbitrum",
                          "arbitrum-sepolia",
                          "robinhood-testnet",
                          "arc-testnet"
                        ]
                      },
                      "payerAddress": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{40}$"
                      }
                    },
                    "required": ["asset", "chain"]
                  },
                  "executionPath": {
                    "type": "string",
                    "enum": ["deposit-match", "on-chain-contract", "x402"]
                  },
                  "amount": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "string"
                      },
                      "asset": {
                        "type": "string",
                        "enum": [
                          "IDR",
                          "USD",
                          "SGD",
                          "THB",
                          "MYR",
                          "USDC",
                          "USDT",
                          "EURC",
                          "ETH",
                          "BTC"
                        ]
                      }
                    },
                    "required": ["amount", "asset"]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment intent created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIntent"
                }
              }
            }
          },
          "404": {
            "description": "No invoice with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/checkout/{id}": {
      "get": {
        "tags": ["Unversioned"],
        "operationId": "getCheckoutPage",
        "summary": "Hosted checkout page",
        "description": "Intentionally unversioned. Renders the hosted checkout for a payment link. A printed QR and a shared link encode this URL, so a version bump must never move it.",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Checkout HTML page.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "No link with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}